apps/api/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
pydantic-settings>=2.2,<3.0
PyJWT>=2.8,<3.0
passlib[bcrypt]>=1.7,<2.0
email-validator>=2.2,<3.0
httpx>=0.28,<1.0
pytest>=8.0,<9.0
